Moved config.py to oemof_b3.config
